- Store coaching pace values with decimal precision\n- Add community search and multipart community endpoints\n- Cover coaching precision and community search responses with tests

+++ src/main/java/com/neostride/server/coaching/ai/AiCoachingPlanDay.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 public record AiCoachingPlanDay(
 		@JsonProperty("plan_date") LocalDate planDate,
 		@JsonProperty("day_distance_km") BigDecimal dayDistanceKm,
-		@JsonProperty("day_pace_min_per_km") Integer dayPaceMinPerKm,
+		@JsonProperty("day_pace_min_per_km") BigDecimal dayPaceMinPerKm,
 		@JsonProperty("description") String description
 ) {
 }
